test(desktop): load app center state source without JSX
