+++ apps/desktop/src/renderer/src/features/workspace-app-center/services/internal/workspaceAppCenterContribution.tsx
@@ -1,7 +1,6 @@
 import { createElement, useEffect, useRef, type ReactNode } from "react";
 import {
   type BrowserNodeNavigationPolicy,
-  type BrowserNodeRuntimeState,
   type BrowserNodeFeature
 } from "@tutti-os/browser-node";
 import { BrowserNode } from "@tutti-os/browser-node/react";
@@ -19,8 +18,6 @@ import { createAppCenterI18nRuntime } from "@tutti-os/workspace-app-center/i18n"
 import type {
   WorkbenchContribution,
   WorkbenchHostDockEntry,
-  WorkbenchHostExternalStateLookupInput,
-  WorkbenchHostExternalStateSource,
   WorkbenchHostNodeBodyContext,
   WorkbenchHostNodeHeaderContext,
   WorkbenchHostNodeDefinition
@@ -39,6 +36,7 @@ import type {
   WorkspaceAppCenterViewState
 } from "@tutti-os/workspace-app-center";
 import { createWorkspaceAppCenterOpenedLease } from "./workspaceAppCenterAnalytics.ts";
+import { createWorkspaceAppWebviewExternalStateSource } from "./workspaceAppCenterExternalState.ts";
 import {
   WorkspaceAppCenterInlineAppBody,
   workspaceAppBrowserPartitionPrefix
@@ -63,7 +61,6 @@ import {
 import {
   findWorkspaceApp,
   readWorkspaceAppIdFromInstanceId,
-  readWorkspaceAppIdFromNodeId,
   resolveWorkspaceAppCenterLaunchRequest,
   resolveWorkspaceAppDisplayName,
   workspaceAppCenterNodeID,
@@ -722,149 +719,6 @@ function WorkspaceAppWebviewLoadingState({
       </div>
     </div>
   );
-}
-
-type WorkspaceAppCenterExternalNodeState =
-  | WorkspaceAppCenterViewState
-  | WorkspaceAppWebviewExternalState
-  | null;
-
-export function createWorkspaceAppWebviewExternalStateSource(input: {
-  appCenterService: IWorkspaceAppCenterService;
-  runtimeStore: {
-    getSnapshot(): Record<string, BrowserNodeRuntimeState | undefined>;
-    subscribe(listener: () => void): () => void;
-  };
-}): WorkbenchHostExternalStateSource<
-  WorkspaceAppCenterExternalNodeState,
-  null
-> {
-  const appCenterViewStateByWorkspaceId = new Map<
-    string,
-    WorkspaceAppCenterViewState
-  >();
-  const appWebviewStateByNodeId = new Map<
-    string,
-    WorkspaceAppWebviewExternalState
-  >();
-
-  return {
-    getNodeState(request) {
-      if (request.typeId === workspaceAppCenterNodeID) {
-        return reuseWorkspaceAppCenterViewState(
-          appCenterViewStateByWorkspaceId,
-          request.workspaceId,
-          input.appCenterService.getViewState(request.workspaceId)
-        );
-      }
-      return reuseWorkspaceAppWebviewExternalState(
-        appWebviewStateByNodeId,
-        request.nodeId,
-        readWorkspaceAppExternalState(input, request)
-      );
-    },
-    getSnapshotNodeState(request) {
-      if (request.typeId === workspaceAppCenterNodeID) {
-        return reuseWorkspaceAppCenterViewState(
-          appCenterViewStateByWorkspaceId,
-          request.workspaceId,
-          input.appCenterService.getViewState(request.workspaceId)
-        );
-      }
-      return reuseWorkspaceAppWebviewExternalState(
-        appWebviewStateByNodeId,
-        request.nodeId,
-        readWorkspaceAppExternalState(input, request)
-      );
-    },
-    getWorkspaceState() {
-      return null;
-    },
-    subscribeNodeState(request, listener) {
-      if (
-        request.typeId !== workspaceAppCenterNodeID &&
-        request.typeId !== workspaceAppWebviewTypeID
-      ) {
-        return () => {};
-      }
-      const unsubscribeRuntime = input.runtimeStore.subscribe(listener);
-      const unsubscribeAppCenter = input.appCenterService.subscribe(listener);
-      return () => {
-        unsubscribeRuntime();
-        unsubscribeAppCenter();
-      };
-    }
-  };
-}
-
-function reuseWorkspaceAppCenterViewState(
-  stateByWorkspaceId: Map<string, WorkspaceAppCenterViewState>,
-  workspaceId: string,
-  next: WorkspaceAppCenterViewState
-): WorkspaceAppCenterViewState {
-  const previous = stateByWorkspaceId.get(workspaceId);
-  if (
-    previous?.activeAppTab === next.activeAppTab &&
-    previous.openAppId === next.openAppId &&
-    previous.openAppIds?.length === next.openAppIds?.length &&
-    previous.openAppIds?.every(
-      (appId, index) => appId === next.openAppIds?.[index]
-    )
-  ) {
-    return previous;
-  }
-  stateByWorkspaceId.set(workspaceId, next);
-  return next;
-}
-
-function reuseWorkspaceAppWebviewExternalState(
-  stateByNodeId: Map<string, WorkspaceAppWebviewExternalState>,
-  nodeId: string,
-  next: WorkspaceAppWebviewExternalState | null
-): WorkspaceAppWebviewExternalState | null {
-  const previous = stateByNodeId.get(nodeId) ?? null;
-  if (!next) {
-    stateByNodeId.delete(nodeId);
-    return null;
-  }
-  if (previous?.title === next.title && previous.url === next.url) {
-    return previous;
-  }
-  stateByNodeId.set(nodeId, next);
-  return next;
-}
-
-function readWorkspaceAppExternalState(
-  input: {
-    appCenterService: IWorkspaceAppCenterService;
-    runtimeStore: {
-      getSnapshot(): Record<string, BrowserNodeRuntimeState | undefined>;
-    };
-  },
-  request: WorkbenchHostExternalStateLookupInput
-): WorkspaceAppWebviewExternalState | null {
-  if (request.typeId !== workspaceAppWebviewTypeID) {
-    return null;
-  }
-  const runtime = input.runtimeStore.getSnapshot()[request.nodeId];
-  const runtimeUrl = runtime?.url?.trim();
-  if (runtimeUrl) {
-    return {
-      title: runtime?.title?.trim() || null,
-      url: runtimeUrl
-    };
-  }
-
-  const appId =
-    readWorkspaceAppIdFromNodeId(request.nodeId) ??
-    readWorkspaceAppIdFromInstanceId(request.instanceId);
-  const app = appId ? findWorkspaceApp(input.appCenterService, appId) : null;
-  return app?.launchUrl
-    ? {
-        title: resolveWorkspaceAppDisplayName(app),
-        url: app.launchUrl
-      }
-    : null;
 }
 
 function resolveWorkspaceAppWebviewNavigationPolicy(input: {
